refactor: MyChatSection to move to inside screens chat sections
